Remove follows page and refactor profile component(s)
