+++ fe1-web/src/core/navigation/typing/SocialParamList.ts
@@ -7,6 +7,5 @@ import { SocialSearchParamList } from './SocialSearchParamList';
 export type SocialParamList = {
   [STRINGS.social_media_navigation_tab_home]: undefined;
   [STRINGS.social_media_navigation_tab_search]: NavigatorScreenParams<SocialSearchParamList>;
-  [STRINGS.social_media_navigation_tab_follows]: undefined;
   [STRINGS.social_media_navigation_tab_profile]: undefined;
 };
