+++ fe1-web/src/features/social/screens/index.ts
@@ -1,4 +1,3 @@
-export { default as SocialFollows } from './SocialFollows';
 export { default as SocialHome } from './SocialHome';
 export { default as SocialProfile } from './SocialProfile';
 export { default as SocialSearch } from './SocialSearch';
